Move scalar unpack functions to their own file
