+++ cpp/src/arrow/CMakeLists.txt
@@ -490,6 +490,7 @@ set(ARROW_UTIL_SRCS
     util/bitmap_builders.cc
     util/bitmap_ops.cc
     util/bpacking.cc
+    util/bpacking_scalar.cc
     util/byte_size.cc
     util/byte_stream_split_internal.cc
     util/cancel.cc
