+++ cpp/src/arrow/util/bpacking_benchmark.cc
@@ -22,6 +22,7 @@
 
 #include "arrow/testing/util.h"
 #include "arrow/util/bpacking_internal.h"
+#include "arrow/util/bpacking_scalar_internal.h"
 
 #if defined(ARROW_HAVE_RUNTIME_AVX2)
 #  include "arrow/util/bpacking_avx2_internal.h"
