+++ cpp/src/arrow/util/bpacking_scalar.cc
@@ -15,23 +15,22 @@
 // specific language governing permissions and limitations
 // under the License.
 
-#pragma once
-
-#include "arrow/util/visibility.h"
-
-#include <cstdint>
+#include "arrow/util/bpacking_dispatch_internal.h"
+#include "arrow/util/bpacking_scalar_generated_internal.h"
+#include "arrow/util/bpacking_scalar_internal.h"
 
 namespace arrow::internal {
 
-#if defined(ARROW_HAVE_SSE4_2)
-/// The 32 bit unpacking with SSE 4.2
-ARROW_EXPORT int unpack32_sse4_2(const uint8_t* in, uint32_t* out, int batch_size,
-                                 int num_bits);
-#endif
+int unpack16_scalar(const uint8_t* in, uint16_t* out, int batch_size, int num_bits) {
+  return unpack_jump16<ScalarUnpackerForWidth>(in, out, batch_size, num_bits);
+}
 
-ARROW_EXPORT
-int unpack32(const uint8_t* in, uint32_t* out, int batch_size, int num_bits);
-ARROW_EXPORT
-int unpack64(const uint8_t* in, uint64_t* out, int batch_size, int num_bits);
+int unpack32_scalar(const uint8_t* in, uint32_t* out, int batch_size, int num_bits) {
+  return unpack_jump32<ScalarUnpackerForWidth>(in, out, batch_size, num_bits);
+}
+
+int unpack64_scalar(const uint8_t* in, uint64_t* out, int batch_size, int num_bits) {
+  return unpack_jump64<ScalarUnpackerForWidth>(in, out, batch_size, num_bits);
+}
 
 }  // namespace arrow::internal
