+++ cpp/src/arrow/util/bpacking_test.cc
@@ -24,6 +24,7 @@
 #include "arrow/testing/util.h"
 #include "arrow/util/bit_stream_utils_internal.h"
 #include "arrow/util/bpacking_internal.h"
+#include "arrow/util/bpacking_scalar_internal.h"
 #include "arrow/util/logging.h"
 
 #if defined(ARROW_HAVE_RUNTIME_AVX2)
